Query 1 Functional Plots, Sliders, near functionality of visibility
Each graph successfully renders based on the spliced query results. The sliders for all plots work properly. The visibility checks for the severity plot of query 1 function, but these same checks are currently bugged for the other plots. Improvements to these and to the design/layout are being worked on.