Added separate pages for each query, altered how line visibility worked, added query 5 graphs
Now, there will be a home page jupyter notebook file, that will have a description (pending) of the entire project and buttons that link to each query's page. Each query page will have graphs associated with the query, descriptions of each graph and its results, and a button that links back to the home page. Query 5 has been added as well.